Remove node modules, add test framework
