+++ package.json
@@ -10,11 +10,14 @@
     "node-fetch": "^2.6.1"
   },
   "devDependencies": {
-    "@vercel/ncc": "^0.25.1"
+    "@vercel/ncc": "^0.25.1",
+    "@zeit/ncc": "^0.22.3",
+    "jest": "^26.6.3"
   },
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "ncc build index.js --minify",
+    "test": "jest"
   },
   "repository": {
     "type": "git",
